QQuickListView: implement support for reusing items
This patch will implement delegate item recycling in
ListView. The API will be the same as used in TableView, except
that it will be off by default since the behaviour of a delegate
that is reused is not compatible with legacy applications
that was not written with this in mind.

Most importantly:

- Component.onCompleted will only be called on a delegate the
first time it's created, and never when it's reused.

- Any application specific properties in the delegate (that is, not
model roles, index, etc) will not be cleared or updated when an
item is reused. The application must do this manually upon
receiving e.g the ListView.onReused signal.

[ChangeLog][ListView] ListView now has support for reusing delegate
items. This can be switched on by setting ListView.reuseItems: true

Change-Id: I68cc8300b050e4a1f89feebb1d31a2fd9189c793