QQmlInstanceModel::ReleaseFlags: add enum value 'Pooled'
Now that we're about to add support for delegate item
recycling in ListView, we need to add an extra enum
value to the ReleaseFlags. This will be needed later so
that ListView can distinguish between items that are
still referenced and visible in the viewport, and items
that not referenced, but at the same time, still
alive in the pool.

Change-Id: I4a1110b6b43ba109ccd160d22010569dd5410829